TOOLS: create `ReplUtil`
